Rename jobs_manager to match class name
